Return CGFloats from the profile size calculators

+++ Sources/Controllers/Profile/Calculators/ProfileViewsSizeCalculator.swift
@@ -7,8 +7,8 @@ import FutureKit
 
 public class ProfileViewsSizeCalculator: NSObject {
 
-    public func calculate(item: StreamCellItem) -> Future<Int> {
-        let promise = Promise<Int>()
+    public func calculate(item: StreamCellItem) -> Future<CGFloat> {
+        let promise = Promise<CGFloat>()
         promise.completeWithSuccess(122)
         return promise.future
     }
